:recycle: [태진] refactor: MessageCard, EditCard를 Card로 통합
