+++ src/pages/Paper/components/Card/CardStyled.jsx
