--- src/pages/Paper/components/EditCard/index.js
@@ -1,1 +1,0 @@
-export { default } from './EditCard';
